Fixed sudden fps drop on summons pathfinding
Fixed a sudden fps drop on summons. This was caused because of summons nav agent recalculating the path indefinitely when an idle pos was not reacheble. It was fixed by finding a new idle pos that is near to the player, and if no valid idle pos is found then the idle pos is player position which should always be valid.